- Ajustes

+++ maze.hpp
@@ -11,20 +11,20 @@ namespace Maze
     };
 
     struct Line {
-        int x0, y0;
-        int x1, y1;
+        float x0, y0;
+        float x1, y1;
     };
 
     struct Rect {
-        int x, y;
-        int width, height;
+        float x, y;
+        float width, height;
     };
 
     auto make(size_t rows, size_t columns)->std::vector<std::vector<Tile>>;
 
     auto print(const std::vector<std::vector<Tile>>& matrix) -> void;
 
-    auto lines(const std::vector<std::vector<Tile>>& matrix, int height, int width)->std::vector<Line>;
+    auto lines(const std::vector<std::vector<Tile>>& matrix, float height, float width)->std::vector<Line>;
 
-    auto rectangles(const std::vector<std::vector<Tile>>& matrix, int x, int y, int height, int width, int thickness)->std::vector<Rect>;
+    auto rectangles(const std::vector<std::vector<Tile>>& matrix, float x, float y, float height, float width, float thickness)->std::vector<Rect>;
 }
